# backend/openapi.yaml
...
openapi: 3.0.3
info:
  title: CarbonManager On-Chain API
  version: 1.0.0
servers:
  - url: http://localhost:5001
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
  /onchain/emissions/{emission_id}:
    post:
      summary: Create/refresh a pending on-chain job for an emission
      tags: [Blockchain]
      parameters:
        - in: path
          name: emission_id
          required: true
          schema: { type: integer, minimum: 1 }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  emission_id: { type: integer }
                  status: { type: string, enum: [pending] }
                  payload: { type: object }
        "404":
          description: Emission not found
    get:
      summary: Get on-chain status for an emission
      tags: [Blockchain]
      parameters:
        - in: path
          name: emission_id
          required: true
          schema: { type: integer, minimum: 1 }
      responses:
        "200":
          description: Status record
          content:
            application/json:
              schema:
                type: object
                properties:
                  emission_id: { type: integer }
                  status: { type: string, enum: [pending,submitted,confirmed,failed] }
                  tx_hash: { type: string, nullable: true }
                  payload_json: { type: object }
                  error_msg: { type: string, nullable: true }
        "404":
          description: No on-chain record
...
  /products/{product_id}/emissions:
    get:
      summary: Get all emissions under a product
      tags: [Product Emissions]
      security:
        - BearerAuth: []
      parameters:
        - name: product_type_id
          in: path
          required: true
          schema: { type: string, minimum: 1 }
        - name: product_id
          in: path
          required: true
          schema: { type: string, minimum: 1 }
      responses:
        "200":
          description: List of emissions
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    amount:
                      type: number
                      format: float
                    timestamp:
                      type: string
                      format: date-time
        "404":
          description: Product not found
    post:
      summary: Create a new emission under a product
      tags: [Product Emissions]
      security:     
        - BearerAuth: []
      parameters:
        - name: product_type_id
          in: path
          required: true
          schema: { type: string, minimum: 1 }
        - name: product_id
          in: path
          required: true
          schema: { type: string, minimum: 1 }
      requestBody:    
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                stage_id:
                  type: string 
                  example: raw
                factor_id:
                  type: integer
                  example: 1
                material:
                  type: string
                  example: 瓶裝水(2000ml，PET包裝)
                unit:
                  type: string
                  example: 瓶
                amount:
                  type: number
                  format: float
                  example: 2
      responses:
        "201":
          description: Emission created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  emission_id:
                    type: string
                  status_message:
                    type: string
...
  /emissions:
    get:
      summary: Get all emissions from one organization, across all product types and products
      tags: [Emissions]
      security:
        - BearerAuth: []
      responses:
        "200":
          description: List of all emissions
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    emission_id:
                      type: string
                    emission_name:
                      type: string
                    stage_id:
                      type: string
                    step_id:
                      type: string
                    tag_id:
                      type: string
                    factor_id:
                      type: integer
                    quantity:
                      type: number
                      format: float
                    emission_amount:
                      type: number
                      format: float
                    created_at:
                      type: string
                    created_by:
                      type: string
  /emissions/{emission_id}:
    get:
      summary: Get an emission by ID
      tags: [Emissions]
      security:
        - BearerAuth: []
      parameters:
        - name: emission_id
          in: path
          required: true
          schema: { type: string, minimum: 1 }     
      responses:
        "200":
          description: Emission details
          content:
            application/json:
              schema:   
                type: object
                properties:
                  emission_id:
                    type: string
                  emission_name:
                    type: string
                  stage_id:
                    type: string
                  step_id:
                    type: string
                  tag_id:
                    type: string
                  factor_id:
                    type: integer
                  quantity:
                    type: number
                    format: float
                  emission_amount:
                    type: number
                    format: float
                  unit_target_amount:
                    type: number
                  created_at:
                    type: string
                  created_by:
                    type: string
        "404":
          description: Emission not found
    put:
      summary: Update an emission by ID
      tags: [Emissions]
      security:
        - BearerAuth: []        
      parameters:
        - name: emission_id
          in: path
          required: true
          schema: { type: string, minimum: 1 }
      requestBody:      
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                new amount:
                  type: number
                  format: float
                  example: 23.45
      responses:
        "200":
          description: Emission updated successfully
        "400":
          description: Invalid input
        "401":
          description: Unauthorized
        "404":
          description: Emission not found
    delete:
      summary: Delete an emission by ID 
      tags: [Emissions]
      security:
        - BearerAuth: []
      parameters:     
        - name: emission_id
          in: path
          required: true        
          schema: { type: string, minimum: 1 }
      responses:  
        "200":
          description: Emission deleted successfully
          content:
            application/json: 
              schema:
                type: object
                properties:
                  status_message:
                    type: string
        "401":
          description: Unauthorized
        "404":
          description: Emission not found
...